add ports sales

+++ saleservice/src/main/java/br/com/mmmsieto/saleservice/application/core/domain/Sale.java
@@ -9,7 +9,7 @@ public class Sale {
     public Sale() {
     }
 
-    public Sale(Integer productId, Integer userId, BigDecimal price, SaleStatus status, Integer quantity) {
+    public Sale(Long productId, Long userId, BigDecimal price, SaleStatus status, Long quantity) {
         this.productId = productId;
         this.userId = userId;
         this.price = price;
@@ -17,46 +17,46 @@ public class Sale {
         this.quantity = quantity;
     }
 
-    private Integer id;
+    private Long id;
 
-    private Integer productId;
+    private Long productId;
 
-    private Integer userId;
+    private Long userId;
 
     private BigDecimal price;
 
     private SaleStatus status;
 
-    private Integer quantity;
+    private Long quantity;
 
-    public Sale(Integer productId, Integer userId, BigDecimal price, Integer quantity) {
+    public Sale(Long productId, Long userId, BigDecimal price, Long quantity) {
         this.productId = productId;
         this.userId = userId;
         this.price = price;
         this.quantity = quantity;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(Integer productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -76,11 +76,11 @@ public class Sale {
         this.status = status;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
 }
